+++ strawberryfields/backends/gaussianbackend/ops.py
@@ -39,6 +39,7 @@ def chop_in_blocks_vector(v, idtodelete):
     which elements go into vb
     """
     idtokeep = list(set(np.arange(len(v))) - set(idtodelete))
+    idtokeep.sort()
     va = v[idtokeep]
     vb = v[idtodelete]
     return (va, vb)
@@ -51,7 +52,8 @@ def reassemble(A, idtodelete):
     The empty space are filled with zeros (offdiagonal) and ones (diagonals)
     """
     ntot = len(A) + len(idtodelete)
-    ind = set(np.arange(ntot)) - set(idtodelete)
+    ind = list(set(np.arange(ntot)) - set(idtodelete))
+    ind.sort()
     newmat = np.zeros((ntot, ntot))
     for i, i1 in enumerate(ind):
         for j, j1 in enumerate(ind):
@@ -67,7 +69,8 @@ def reassemble_vector(va, idtodelete):
     and everywhere else it puts the entries of va
     """
     ntot = len(va) + len(idtodelete)
-    ind = set(np.arange(ntot)) - set(idtodelete)
+    ind = list(set(np.arange(ntot)) - set(idtodelete))
+    ind.sort()
     newv = np.zeros(ntot)
     for j, j1 in enumerate(ind):
         newv[j1] = va[j]